+++ History.md
@@ -7,6 +7,7 @@
  * Inlined some settings
  * Added build script, which produces nprogress.min.js
  * The CSS class 'nprogress-busy' is no longer added to the document
+ * Made NProgress.status private
 
 ## [v0.2.0] - May 13, 2015